Sync Fork from Upstream Repo

+++ apps/utilities/kill.cpp
@@ -1,3 +1,4 @@
+#include <libio/File.h>
 #include <libsystem/process/Process.h>
 #include <libutils/ArgParse.h>
 #include <libutils/NumberParser.h>
@@ -15,6 +16,14 @@ int kill(int pid)
 
 int killall(String name)
 {
+    IO::File file{"/System/processes", OPEN_READ};
+
+    if (!file.exist())
+    {
+        IO::errln("/System/processes not found");
+        return PROCESS_FAILURE;
+    }
+
     if (name == "neko")
     {
         IO::errln("Don't kill nekos, your are a bad persone!");
@@ -25,7 +34,7 @@ int killall(String name)
         IO::errln("Don't kill cats, you monster!");
     }
 
-    auto processes = Json::parse_file("/System/processes");
+    auto processes = Json::parse(file);
 
     for (size_t i = 0; i < processes.length(); i++)
     {
